+++ ErsatzTV.FFmpeg/RateControlMode.cs
@@ -1,0 +1,11 @@
+using System.Diagnostics.CodeAnalysis;
+
+namespace ErsatzTV.FFmpeg;
+
+[SuppressMessage("ReSharper", "InconsistentNaming")]
+public enum RateControlMode
+{
+    CBR,
+    CQP,
+    VBR
+}